refactor: Move dynamic lists out of Upload & RecipeFilter
