+++ react-frontend/src/components/Upload.tsx
@@ -1,70 +1,7 @@
 import { useState } from "react";
 import api from "../config/axios_config";
 import { useNavigate } from "react-router-dom";
-
-interface ListItemProps {
-    item: string,
-    index: number,
-    onDeleteItem: (index: number) => void,
-    onEditItem: (index: number, editValue: string) => void
-}
-
-interface ListProps {
-    items: string[],
-    onDeleteItem: (index: number) => void,
-    onEditItem: (index: number, editValue: string) => void
-}
-
-function ListItem({item, index, onDeleteItem, onEditItem}: ListItemProps) {
-    const [isEditing, setIsEditing] = useState(false);
-    const [editValue, setEditValue] = useState(item);
-
-    const saveEditHandler = () => {
-        onEditItem(index, editValue);
-        setIsEditing(false);
-    };
-
-    const cancelEditHandler = () => {
-        setIsEditing(false);
-    }
-
-    return (
-        <li>
-            {
-                isEditing ? (<>
-                    <input 
-                        type="text" 
-                        value={editValue} 
-                        onChange={(e) => setEditValue(e.target.value)}
-                    />
-                    <button onClick={saveEditHandler}>Save</button>
-                    <button onClick={cancelEditHandler}>Cancel</button>
-                </>) : (<>
-                    {item}
-                    <button onClick={() => setIsEditing(true)}>Edit</button>
-                    <button onClick={() => onDeleteItem(index)}>Delete</button>
-                </>)
-            }
-        </li>
-    );
-}
-
-function List({items, onEditItem, onDeleteItem}: ListProps) {
-    return (
-        <ul>
-            {
-                items.map((item, index) => (
-                    <ListItem 
-                        key={index}
-                        item={item} index={index} 
-                        onEditItem={onEditItem} 
-                        onDeleteItem={onDeleteItem}
-                    />
-                ))
-            }
-        </ul>
-    );
-}
+import DynamicEditList from "./DynamicEditList";
 
 function Upload() {
     const navigate = useNavigate();
@@ -168,7 +105,7 @@ function Upload() {
                 <button onClick={addIngredientHandler}>Add</button>
             </>
             <>
-                <List
+                <DynamicEditList
                     items={ingredients}
                     onEditItem={editIngredientHandler}
                     onDeleteItem={deleteIngredientHandler}
@@ -186,7 +123,7 @@ function Upload() {
                 <button onClick={addInstructionHandler}>Add</button>
             </>
             <>
-                <List
+                <DynamicEditList
                     items={instructions}
                     onEditItem={editInstructionHandler}
                     onDeleteItem={deleteInstructionHandler}
